requirements.txt
fastapi
uvicorn
google-genai
sentence-transformers
datasets
pyarrow
rank-bm25
torch
transformers
chromadb
openai
ddgs
beautifulsoup4
langchain-text-splitters
selenium
seaborn
docling
timeout-function-decorator